Fix meetingroom create url parameters
For bbb-recording-ready-url:
Apparently, this was earlier mis-typed as bn-recording-ready-url in the
BBB documentation and fixed some time ago
(https://github.com/bigbluebutton/bigbluebutton/issues/8216).

For now, I just changed the name of the URL parameter and not also the
member name in the Meetingroom struct, because that would require some
conversion code between the two variants, when loading older
meetingrooms from the database.

For endCallbackUrl:
Also, I changed the camelCase of endCallbackUrl so it is in line with
what the API docs say at
https://docs.bigbluebutton.org/dev/api.html#end-meeting-callback-url.